# Narrow four utf16 chars to ascii and write to buffer