I decided to split up my single PR into multiple ones so I can isolate CI bugs better. I believe this change should be good to go in. 

In the process of splitting, I eventually found that the ARM64 CI legs seem to be broken currently. I opened https://github.com/dotnet/runtime/pull/39582 to test that claim. 

Perf results:
### Before and After Arm64 optimizations
calope@calopearm:~/performance/src/tools/ResultsComparer$ dotnet run --base ~/pgovind_before/ --diff ~/pgovind_after/ --threshold 0.01%
summary:
better: 1, geomean: 1.244
total diff: 1
No Slower results for the provided threshold = 0.01% and noise filter = 0.3ns.
| Faster                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Experimental.Perf.Intrinsics |      1.24 |         10433.00 |          8384.06 |         |
No file given